Inject properties defined in active profiles
Properties defined in profiles which are activated on the command line
(with the -P switch) should be injected into the effective pom as these
properties may be used in pom elements that remain in the flattened pom.

A use case for this is when the scm tag element is set to a property
value that is only defined if a certain profile is enabled.