Fix #26: Resolve multi-module projects
Adds the projects in the reactor to FlattenModelResolver. This is
necessary for multi-module projects where one module imports another as
a bom as this module may not be available in the local repository.

The changes in this commit rely on MavenSession.getAllProjects() to
return the unconstrained list of projects in the reactor. This is not
the case in Maven versions 3.3.1, 3.3.3 and 3.3.9. A fix has been
submitted to Maven (see https://issues.apache.org/jira/browse/MNG-6173).
It is included in Maven starting with 3.5.0-alpha-1. Therefore, the
dependencies of the flatten-maven-plugin are updated to require Maven
3.5.0.

As Travis CI only has Maven 3.2.5, this causes the integration tests for
the flatten-maven-plugin to fail. This commit adds the maven-wrapper [1]
to the project and its Travis CI build configuration. The wrapper script
automatically downloads Maven 3.5.0 and uses it for compilation and test
execution.

[1] https://github.com/takari/maven-wrapper