Inject properties defined in active profiles (#31)
* Inject properties defined in active profiles

Properties defined in profiles which are activated on the command line
(with the -P switch) should be injected into the effective pom as these
properties may be used in pom elements that remain in the flattened pom.

A use case for this is when the scm tag element is set to a property
value that is only defined if a certain profile is enabled.

* Improve integration tests

Removes unnecessary flatten-maven-plugin declarations from the
profile-with-properties-active and profile-with-properties-inactive
tests.

* Do not abuse version element for integration tests

The version element in the pom should not contain expressions but only a
constant value. This commit changes the integration tests which verify
the properties from profiles are applied correctly, to apply the
property to the license name element instead of the version element.