Make up some years of debt
As part of user reported issue https://issues.apache.org/jira/browse/MRESOLVER-322 and
for upcoming 1.9.5 resolver release went to look what it is about, only to figure out
that this plugin accumulate several terrible lag, bad practice and many other things.

Whenever you see "inappropriate dependency" (ie. using MAT in plugin that declares
itself as a 3.2.5+ plugin), use of deprecated components (ArtifactFactory and
relatives), or doing hoops and loops to get dependnecy graph ONLY to do
something about it, step up and fix it.

These changes makes plugin work as expected from declared "minimum" version
(3.2.5) to current stable (3.9.0) with addition that it will support
split repository and any other modern feature incoming....

Lack of time prevented me to finish this PR to 100%, so there are some
IT failures and some TODOs in code left, but results are:

With Maven 3.2.5:
```
[INFO] Build Summary:
[INFO]   Passed: 72, Failed: 4, Errors: 0, Skipped: 2
[INFO] -------------------------------------------------
[ERROR] The following builds failed:
[ERROR] *  flatten-dependency-all-both-test-and-transitive/pom.xml
[ERROR] *  flatten-dependency-all-provided-children/pom.xml
[ERROR] *  flatten-dependency-all-provided-direct/pom.xml
[ERROR] *  parent-with-version-range/pom.xml
[INFO] -------------------------------------------------
[WARNING] The following builds were skipped:
[WARNING] *  complete-multimodule-parent-pom-cifriendly/pom.xml
[WARNING] *  no-overwrite-3.6.3-later/pom.xml
[INFO] -------------------------------------------------
```

With Maven 3.8.7:
```
[INFO] -------------------------------------------------
[INFO] Build Summary:
[INFO]   Passed: 74, Failed: 3, Errors: 0, Skipped: 1
[INFO] -------------------------------------------------
[ERROR] The following builds failed:
[ERROR] *  flatten-dependency-all-both-test-and-transitive/pom.xml
[ERROR] *  flatten-dependency-all-provided-children/pom.xml
[ERROR] *  flatten-dependency-all-provided-direct/pom.xml
[INFO] -------------------------------------------------
[WARNING] The following builds were skipped:
[WARNING] *  no-overwrite-3.6.2-before/pom.xml
[INFO] -------------------------------------------------
```

With Maven 3.9.0:
```
[INFO] -------------------------------------------------
[INFO] Build Summary:
[INFO]   Passed: 74, Failed: 3, Errors: 0, Skipped: 1
[INFO] -------------------------------------------------
[ERROR] The following builds failed:
[ERROR] *  flatten-dependency-all-both-test-and-transitive/pom.xml
[ERROR] *  flatten-dependency-all-provided-children/pom.xml
[ERROR] *  flatten-dependency-all-provided-direct/pom.xml
[INFO] -------------------------------------------------
[WARNING] The following builds were skipped:
[WARNING] *  no-overwrite-3.6.2-before/pom.xml
[INFO] -------------------------------------------------
```