Add skip flags (#385)
* Add skip flags

This introduces 3 flags:
- `flatten.skip`: skips the entire plugin regardless of goal
- `flatten.clean.skip`: skips the clean goal
- `flatten.flatten.skip`: skips the flatten goal

This means the user has control over what parts are skipped, and the
flags are descriptive. The `flatten.` prefix is in line with other
parameters like `flatten.mode` and `flatten.dependency.mode`.

* Test skip flags

This adds 3 ITs for the skip flags:
- to verify `skip` -> should not create/remove flattened pom
- to verify `skipFlatten` -> should not create flattened pom
- to verify `skipClean` -> should not remove flattened pom

* Prepare for 1.6.0 release

* Annotate skip flags as since 1.6.0

This is the release in which they'll be first available.